+++ src/if.rs
@@ -1,126 +1,104 @@
-/*
-Rust import system is so stupid that I can't import anything MY OWN modules in this particular project tree.
-Trying to use mod or use in any way does not help.
+use crate::global::{index, set_index};
 
-The only solution that I found as of now is to create a directory named "IF" and then creating symbolic links to end.rs and helpful.rs.
-Then, I have to add "#![allow(clippy::duplicate_mod)]" to "rush.rs" to tell the compiler, that I want to ignore the fact, 
-that some modules are imported multiple times (like... WHAT???)
+pub fn logic(mut buf: Vec<String>) -> Result<bool, String> {
+    /*
+    if FAIL:thing1 -with-arg -with-another-arg and OK:thing2 or OUT:thing3 and $variable == 10 do
+        say "I'm a bunch of words on your screen"
+    end
 
-It's just broken or I am sick and I can't read the docs properly to find the correct solution.
-TODO
-*/
-
-use std::collections::HashMap;
-use std::process;
-use carrot_libs::args;
-mod helpful;
-use helpful::*;
-
-// This command uses it's own set of splitting words
-const SPLIT_COMMANDS:[&str;8] = ["if", "elseif", "else", "and", "or", "not", "do", "end"];
-
-/*
-While working with this code, it might be usefull for you to get used to some of my own terminology.
-I don't know if names of parts of IF statements are standarized somewhere or something like that but nevermind...
-
-Just know that many of the functions, variables, comments and lot's of other stuff in the code below uses some magical
-words because I have to name them somehow.
-
-The whole code below is an example of an IF statement. Everything from "IF" to "END" is an IF statement.
-
--- This is an "IF" keyword. It starts this particular type of logic.
--- "IF", "ELSEIF" and "ELSE" are called "super operators".
-|
-|   -- Everything between "IF" and "DO" is a "comparison statement"
-|   |
-|   |        -- This is an "OR" keyword. Both "OR" and "AND" (not used in this example) keywords are called "operators".
-|   |        |
-|   |        |             -- This is a "DO" keyword. It's also called a "summarizer" because it sums up all the return codes
-|   |        |             -- from commands in a "comparison statement". If everything went fine (returned a success) - execute a "task"
-|   |        |             -- if not (there is at least one command that returned a failure) - skip a "task" (I'll explain this in a moment)
-|   |        |             -- and jump to "jump spot" (Will explain this too).
-|   |        |             |
-if cmp 1 = 4 or cmp 1 = 1 do
-    say "Equal to one!"   -- This is a "task"
-    say "or to four!"     -- This also is a "task". Like everything between "DO" and "ELSEIF" or "ELSE"
-
-  -- As you can see, the logic is simple. If everything inside comparison statement succeeds, we do the "task",
-  -- if not, we skip the "task" and jump to the next "ELSEIF", "ELSE" or "END".
-  -- This is why those three keywords are often called a "jump spot".
-  |
-elseif cmp 1 = 2 or cmp 1 = 3 do
-    say "Equal to two or three!"  -- Another "task"
-else                              -- Another "jump spot". "ELSE" does have a "comparison statement" nor "DO" keyword
-    say "I don't get it"          -- Task once again
-end                               -- And a final jump spot.
-*/
-
-fn main() {
-    let args = args::args();
-    let opts = args::opts();
-    let (swcs, vals) = args::swcs();
-
-    // Refuse to run when switches have been passed
-    if ! swcs.is_empty() || ! vals.is_empty() {
-        eprintln!("This program does not support any switches and values!");
-        process::exit(1);
-    };
-
-    if opts.is_empty() {
-        eprintln!("The \"IF\" statement requires some arguments!");
-        process::exit(1);
-    }
+    Possible keyword types are:
+    OK - Execute a command name with 0 it's exit code is 0 or 1 if anything else
+    FAIL - Reversed version of OK
+    CODE - Replace command name
+    OUT - Replace command with it's stdout output
+    ERR - Replace command with it's stderr output
+    LOGIC - AND/OR/DO
+    COMPARATOR - ==, <, >, >=, etc.
+    VAR - Replace variable with it's contents
+    NUMVAL - Raw number
+    TXTVAL - Raw text
+    OKVAL - Raw boolean
     
-    // If there is any "IF" in options, that means, that user probably requested IF multiple times
-    // you can't do that while working with IFs
-    if opts.contains(&"if".to_string()) {
-        eprintln!("SYNTAX ERROR! Repeated \"IF\" operator inside of an IF statement!");
-        process::exit(1);
-    }
-    // If these keywords are not present - The syntax is surelly incorrect
-    if ! opts.contains(&"do".to_string()) {
-        eprintln!("SYNTAX ERROR! Missing \"DO\" operator inside of an IF statement!");
-        process::exit(1);
-    }
-    if ! opts.contains(&"end".to_string()) {
-        //dbg!(&opts);
-        eprintln!("SYNTAX ERROR! Missing \"END\" operator inside of an IF statement!");
-        process::exit(1);
-    }
+    1. Make a hash map of words from IF/ELSEIF/ELSE until DO.
+       Key is a type of the thing and value is the... value.
+        > wordlist = [FAIL:thing1 -with-arg -with-another-arg, LOGIC:AND, OK:thing2, LOGIC:OR, OUT:thing3, LOGIC:AND, VAR:variable, COMPARATOR:==, NUMVAR:10, LOGIC:DO]
     
-    // Protect from writing "if do". The comparison statement cannot be empty.
-    if opts[0] == "do" {
-        eprintln!("SYNTAX ERROR! There is nothing between keywords \"IF\" and \"DO\"!");
-        process::exit(1);
-    }
+    2. Iterate through wordlist
 
-    // Split all arguments by splitting keywords
-    let all_commands = match helpful::split_commands(args.clone(), SPLIT_COMMANDS.to_vec(), false) {
-        Err(e) => { eprintln!("IF OPERATOR FAILED! {e}!"); process::exit(1) },
-        Ok(e) => e,
-    };
-    // Collect exit statuses here
-    let mut returns: HashMap<usize, bool> = HashMap::new();
+    3. Look up for the first word in a list. It may be anything but LOGIC and COMPARATOR.
+       We do not accept LOGICs nor COMPARATORs at this moment because we want to prevent the user from typing something
+       like "if and thing1 do ..." or "if == thing1 and thing2 do ..."
+
+    4. At this point, we need a left comparable object.
+       This has to be a key of type OK, FAIL, CODE, OUT, ERR, VAL, OKVAL, NUMVAL or TXTVAL.
+       No LOGIC nor COMPARATOR is allowed.
+       If you find a key of type OK:
+        > Execute the command with all of it's arguments. Quit from IF block if it does not exist.
+        > Replace it's name depending on it's exit code.
+            If command succeeded (returned exit code 0), set it to [OKVAL:1]
+            If command failed (returned any other exit code), save it as [OKVAL:0]
+       If you find a key of type FAIL:
+        > Do everything like in the case of OK but in reverse.
+       If you find a key of type CODE, OUT or ERR:
+        > Execute the command with all of it's arguments. Quit from IF block if it does not exist.
+        > Replace it's name with the value returned from stdout, stderr or an exit code.
+        > Check if it's a number or a text
+        > Replace it.
+            So now something like "[CODE:thing1 -with-arg -with-another-arg]" may change to "[NUMVAL:0]"
+            and something like "[OUT:thing1 -with-arg -with-another-arg]" may change to "[TXTVAL:Output from thing1]"
+       If you find a key of type VAR:
+        > Replace it's value with variable contents
+        > Check if it's a boolean, number or a text
+        > Replace it like in example below.
+            If variable is set to 10, "[VAR:variable]" will became "[NUMVAL:10]"
+            if variable is set to "hello", "[VAR:variable]" will became "[TXTVAL:hello]"
+            if variable is set to TRUE, "[VAR:variable]" will became "[OKVAL:1]"
+            if variable is set to FALSE, "[VAR:variable]" will became "[OKVAL:0]"
+            BUT if it's set to "TRUE" or "FALSE", it will became "[TXTVAL:TRUE]" or "[TXTVAL:FALSE]"
+       Leave every OKVAL, NUMVAL and TXTVAL as is.
+    5. After exactly one CODE/OUT/ERR/VAL was translated to OKVAL, NUMVAL or TXTVAL,
+       a COMPARATOR or LOGIC is required.
+       If you find a [LOGIC:DO]:
+        > Check if previous block is of type OKVAL
+        > Add to a global script iteration index a value which is an index of DO. (Do smth like INDEX += position-of-do NOT INDEX = position_of_do).
+        > Set shell_mode as CmpSuccess or CmpFailure based on the value from OKVAL 
+        > End this IF statement
+       If you find a [LOGIC:AND] or [LOGIC:OR]:
+        > IF statement can't be finished yet. Report the need for another comparable thing.
+        > Check if previous block is of type OKVAL
+        > At this point, just set if_operation_mode to AND or OR
+       If you find a COMPARATOR:
+        > IF statement can't be finished yet. Report the need for another comparable thing.
+        > Check if previous block is of type NUMVAL or TXTVAL
+        > At this point, just set if_operation_mode to...
+          == and =        EQUAL (acceptable for both types)
+          !=, != and !    DIFFERENT (acceptable for both types)
+          <               LESS (only for NUMVALs)
+          =< or <=        LESS_OR_EQUAL (only for NUMVALs)
+          >               GREATER (only for NUMVALs)
+          >= or =>        GREATER_OR_EQUAL (only for NUMVALs)
+          ~~, =~, ~=, ~   CONTAINS (only for TXTVALs)
+    6. After exactly one COMPARATOR or LOGIC (other than DO), another OK, FAIL, CODE, OUT, ERR, VAL, NUMVAL or TXTVAL is needed
+    7. Repeat step 4 to parse right comparable object of type OK, FAIL, CODE, OUT, ERR, VAL, NUMVAL or TXTVAL.
     
-    let mut idx = 0;
-    // Protect from writing "do end". The task cannot be empty.
-    let do_position = all_commands.iter().position(|x| x[0] == "do" ).unwrap();
-    if all_commands[do_position+1][0] == "end" {
-        eprintln!("SYNTAX ERROR! There is nothing between keywords \"DO\" and \"END\"!");
-        process::exit(1);
+    8. This was a right comparable object in an LEFT_CMD LOGIC/COMPARATOR RIGHT_CMD block.
+       Now, check type of this and this-2 wordlist element
+       and allow comparing them depending on if_operation_mode value.
+       > if_operation_mode is AND / OR: Both elements must be of type OKVAL
+       > if_operation_mode is EQUAL / DIFFERENT:  Both elements must be of type TXTVAL or NUMVAL
+       > if_operation_mode is LESS, LESS_OR..., GREATER, GREATER_OR...: Both elements must be of type NUMVAL
+       > if_operation_mode is CONTAINS: Both elements must be of type TXTVAL
+       In other case, just skip comparison.
+    10. Replace those three elements with OKVAL:1 if...
+       > if_operation_mode = AND and both comparable objects are set to OKVAL:1
+       > if_operation_mode = OR and at least one comparable object is set to OKVAL:1
+       > if_operation_mode is EQUAL/DIFFERENT/LESS/etc. and they match in type and value
+       Otherwise, replace them with OKVAL:0
+    11. Iterate through wordlist again
+    */
+    //let a = index();
+    for w in buf {
+        println!("{w}");
     }
-
-    let end_position = all_commands.iter().position(|x| x[0] == "end" ).unwrap();
-    // dbg!(&all_commands);
-    while idx != end_position {
-        // This binary can be invoked as "if", "elseif" or "else"
-        // Check how it has been run
-        match all_commands[idx][0].as_str() {
-            "if" | "elseif" => make_comparison(&mut idx, &all_commands, &mut returns, false),
-            "else" => make_comparison(&mut idx, &all_commands, &mut returns, true),
-            a => {
-                unreachable!("Internal error! Logic jumped to unknown super operator: {a}!");
-            }
-        }
-    }
+    Ok(true)
 }